+++ src/app/src/index.tsx
@@ -17,11 +17,15 @@ export function renderApplication() {
   const rootElement = document.getElementById("root");
   const adapter = (Native.getInstance().getDefaultConnector() || "").split(".")[2] || "";
   console.debug(adapter)
-  render(<>
+  render(
+    <>
       <Helmet>
         <body className="bp4-dark" data-adapter={adapter} />
-      </Helmet><App store={store} />
-  </>, rootElement);
+      </Helmet>
+      <App store={store} />
+    </>,
+    rootElement
+  );
 }
 
 renderApplication();